pre merge

+++ src/StopD.sh
@@ -14,6 +14,12 @@ fi
 
 pName="$1"
 
+if [ "$1" = "-h" ]; then
+	ayuda
+	exit 1
+fi
+
+
 for ID in ` ps -C "$pName" -o "pid=" `
 do
 	kill -9 $ID
